refactor storage 代码
