+++ engine/persistent_storage_worker.go
@@ -5,8 +5,6 @@ import (
 	"encoding/binary"
 	"encoding/gob"
 	"github.com/huichen/wukong/types"
-	//"io"
-	//"log"
 	"sync/atomic"
 )
 
@@ -46,42 +44,6 @@ func (engine *Engine) persistentStorageRemoveDocumentWorker(docId uint64, shard 
 	// 从数据库删除该key
 	engine.dbs[shard].Delete(b[0:length])
 }
-
-// func (engine *Engine) persistentStorageInitWorker(shard int) {
-// 	iter, err := engine.dbs[shard].SeekFirst()
-// 	if err == io.EOF {
-// 		engine.persistentStorageInitChannel <- true
-// 		return
-// 	} else if err != nil {
-// 		engine.persistentStorageInitChannel <- true
-// 		log.Fatal("无法遍历数据库")
-// 	}
-
-// 	for {
-// 		key, value, err := iter.Next()
-// 		if err == io.EOF {
-// 			break
-// 		} else if err != nil {
-// 			continue
-// 		}
-
-// 		// 得到docID
-// 		docId, _ := binary.Uvarint(key)
-
-// 		// 得到data
-// 		buf := bytes.NewReader(value)
-// 		dec := gob.NewDecoder(buf)
-// 		var data types.DocumentIndexData
-// 		err = dec.Decode(&data)
-// 		if err != nil {
-// 			continue
-// 		}
-
-// 		// 添加索引
-// 		engine.internalIndexDocument(docId, data)
-// 	}
-// 	engine.persistentStorageInitChannel <- true
-// }
 
 func (engine *Engine) persistentStorageInitWorker(shard int) {
 	engine.dbs[shard].ForEach(func(k, v []byte) error {
